requirements.txt
firebase-admin>=6.2.0
google-cloud-datastore>=2.19.0
